Implement brick functionality (#1)
* Updated Bricks

Added functionality to the bricks. Namely, they get destroyed when damaged enough and can emit a specified number of points

* add spawning of bricks

implemented code that spawns the bricks in the game (to be followed up by bricks responding to ball)

* added signals to ball

Added two new signals to the ball scene: the first signal is to inflict damage to bricks, while the second one is to indicate when the player misses the ball and loses a life

* Update Collision Logic

Update the game to use collision masks and layers properly. Also figured out how to get bricks to properly interact with ball

* finished implementing brick functionality

Points now increase when bricks are destroyed