Dev (#3)
* create game environment

create the basic environment, including barriers to stop the ball from moving out of bounds, the player, as well as a ball and some bricks (last two need to be added)

* setting up game

setting up the base of the breakout game

* Implement brick functionality (#1)

* Updated Bricks

Added functionality to the bricks. Namely, they get destroyed when damaged enough and can emit a specified number of points

* add spawning of bricks

implemented code that spawns the bricks in the game (to be followed up by bricks responding to ball)

* added signals to ball

Added two new signals to the ball scene: the first signal is to inflict damage to bricks, while the second one is to indicate when the player misses the ball and loses a life

* Update Collision Logic

Update the game to use collision masks and layers properly. Also figured out how to get bricks to properly interact with ball

* finished implementing brick functionality

Points now increase when bricks are destroyed

* Finalize game (#2)

* implement game loop

Have the game reset paddle and ball position each time a player loses a life. Continues until last life is lost, which removes the ball and stops the game

* add UI

add basic UI showing player's score and lives

* Finish HUD

Have the hud update in time with the game

* Add increasing speed

Cause the ball to increase in speed whenever a brick is broken